Split sensors into submodule and different files
