+++ piweather/sensors/__init__.py
@@ -1,0 +1,4 @@
+from piweather.sensors.Sensor import Sensor  # noqa
+from piweather.sensors.Dummy import Dummy  # noqa
+from piweather.sensors.DS18x20 import DS18x20  # noqa
+from piweather.sensors.A100R import A100R  # noqa
